Refactor: cleaner code, naming
Generally cleaning things up, such as not passing the full list of choices to `updateParent` - just the actual parent tree. (And now it returns that parent tree, not a list).

There are some situations where Answer types are not relevant, for example in `isChildOf` - a tree cannot logically be the child of a tree of type Answer, only Question. But these cases have to be handled anyway.

Being a general function, `listWithoutLast` has been moved to a Utils module.